+++ index.js
@@ -1,5 +1,6 @@
 "use strict"
 
 const TreeProgram = require("./src/TreeProgram.js")
+TreeProgram.NonTerminalNode = require("./src/TreeNonTerminalNode.js")
 
 module.exports = TreeProgram