+++ releaseNotes.md
@@ -1,3 +1,8 @@
+8.2.0 / 2017-10-26
+==================
+- New: export TreeNonTerminalNode class
+- New: support for relative paths in grammar files
+
 8.1.0 / 2017-10-25
 ==================
 - Breaking: renamed fixedWidthTable method to toTable and changed default to left aligned.